UserObj Func adding : delete,select represent obj

+++ src/main/java/com/mindaces/mindaces/api/ValidCheck.java
@@ -66,7 +66,7 @@ public class ValidCheck
     {
         String title = userObjDto.getObjTitle();
         String content = userObjDto.getObjContent();
-        Integer day = userObjDto.getObjDay();
+        Long day = userObjDto.getObjDay();
         if(title.length() < 2 || title.length() > 20){
             return false;
         }else if(content.length() < 2 || content.length() > 40) {
